Extract auth logic to separate module
