+++ auth/authentication.go
@@ -1,4 +1,4 @@
-package controllers
+package auth
 
 import (
 	"fmt"
@@ -10,7 +10,10 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-func signToken(user models.User) (time.Time, string, error) {
+// https://github.com/hay-i/chronologger/issues/35
+var SecretKey = "my_secret"
+
+func SignToken(user models.User) (time.Time, string, error) {
 	expirationTime := time.Now().Add(1 * time.Hour)
 	claims := &jwt.StandardClaims{
 		Subject:   user.Username,
@@ -22,7 +25,7 @@ func signToken(user models.User) (time.Time, string, error) {
 	return expirationTime, signedToken, err
 }
 
-func setCookie(signedToken string, expiry time.Time, c echo.Context) {
+func SetCookie(signedToken string, expiry time.Time, c echo.Context) {
 	cookie := new(http.Cookie)
 	cookie.Name = "token"
 	cookie.Value = signedToken
@@ -33,7 +36,7 @@ func setCookie(signedToken string, expiry time.Time, c echo.Context) {
 	c.SetCookie(cookie)
 }
 
-func parseToken(tokenString string) (jwt.MapClaims, error) {
+func ParseToken(tokenString string) (jwt.MapClaims, error) {
 	token, _ := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
 		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
 			return nil, fmt.Errorf("Unexpected signing method: %v", token.Header["alg"])
