+++ controllers/home.go
@@ -3,6 +3,7 @@ package controllers
 import (
 	"go.mongodb.org/mongo-driver/mongo"
 
+	"github.com/hay-i/chronologger/auth"
 	"github.com/hay-i/chronologger/components"
 	"github.com/labstack/echo/v4"
 )
@@ -14,7 +15,7 @@ func Home(database *mongo.Database) echo.HandlerFunc {
 		if err != nil {
 			isLoggedIn = false
 		} else {
-			_, err = parseToken(cookie.Value)
+			_, err = auth.ParseToken(cookie.Value)
 
 			if err != nil {
 				isLoggedIn = false
